Clean up and refactor WA test code
