+++ tests/weather_anomaly/basics/test_monthly_total_precipitation.py
@@ -1,0 +1,38 @@
+from pycds.weather_anomaly import \
+    monthly_total_precipitation_with_total_coverage, \
+    monthly_total_precipitation_with_avg_coverage, \
+    MonthlyTotalPrecipitation
+
+
+def check(results, months, variable):
+    assert {r.vars_id for r in results} == {variable.id}
+    assert {r.obs_month.month for r in results} == set(months)
+
+
+def test_monthly_total_precipitation_with_total_coverage_query(
+        obs1_precip_sesh, obs1_months, var_precip_net1_1
+):
+    sesh = obs1_precip_sesh
+    monthly_total_precip = \
+        monthly_total_precipitation_with_total_coverage() \
+            .with_session(sesh).all()
+    check(monthly_total_precip, obs1_months, var_precip_net1_1)
+
+
+def test_monthly_total_precipitation_with_avg_coverage_query(
+        obs1_precip_sesh, obs1_months, var_precip_net1_1
+):
+    sesh = obs1_precip_sesh
+    monthly_total_precip = \
+        monthly_total_precipitation_with_avg_coverage() \
+            .with_session(sesh).all()
+    check(monthly_total_precip, obs1_months, var_precip_net1_1)
+
+
+def test_monthly_total_precipitation_matview(
+        obs1_precip_sesh, obs1_months, var_precip_net1_1
+):
+    sesh = obs1_precip_sesh
+    MonthlyTotalPrecipitation.refresh(sesh)
+    monthly_total_precip = sesh.query(MonthlyTotalPrecipitation).all()
+    check(monthly_total_precip, obs1_months, var_precip_net1_1)
