+++ tests/weather_anomaly/test_wa_monthly_views_common.py
